Use modulepath if symlinks aren't in spec/fixtures
Previously, testing rspec-puppet tests required creating symlinks in the
spec/fixtures directory to any modules that could possibly be included
in the tests. This was tedious and annoying. This commit introduces a
change wherein the user's local Puppet modulepath will be used to search
for an included class in the spec tests - which eliminates the need for
creating the dirty symlinks.