Allow Travisci to monitor the module
Previously, the module didn't have any automated TravisCI testing (and
it also didn't have any spec tests to be AUTOMATED). There's another
open pull request to commit spec tests for the module, and this commit
introduces a .travis.yml file that can be used to automate tests
according to Ruby and Puppet versions.