Extract version number directly from git
This uses `git describe` to extract the version number, automatically, in the
Modulefile - rather than forcing manual updates.

This handles dirty branches and nightly builds, though not as nicely as
I would like: because Puppet is strictly SemVer 1.0.0, we only have release
and pre-release versioning.

So, when a nightly is run we assume that the "trivial" version is bumped, and
mark as a pre-release of that.  This is the smallest step possible and ensures
that a full release will *never* be "less" than a nightly build.

Optimally we would postfix a SemVer 2.0.0 build version, but until that is in
all versions of Puppet in the field ... we are stuck with this.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>