Only use annotated git tags for versioning
I had turned on the option to use normal, unannotated tags as well as the
fully, heavy annotated tags that we actually use to indicate a release.

This broke when Jenkins internally tagged the repo with the build details,
since that light-weight tag was closer than the real version tag.

This drops that extra option to `git describe` to avoid that problem.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>