(#98) Removed version from broker arguments.
This change removes the version element from the broker parameters
because it was actually moved into the req_metadata_hash element. The
same happened to the servers element, in revision 66198f54 merged in
0c5a1c303, that was also moved to req_metadata_hash. The difference is
that the servers element was removed from the broker parameters but the
version element was forgotten.

The change is necessary because it prevents an error from occuring
when calling rz_broker. This currently results in the error: Invalid
parameter version(:version).