Update module for CVE-2013-0269 - Ruby JSON vulnerabilities
CVE-2013-0269 details a potential denial-of-service, and remote object
creation, attack on Ruby JSON parsing gems. While Razor is generally safe
against these, and doesn't really have much by way of security to bypass with
object creation, we should protect against those.

This updates the module to ensure that a secure version of the JSON gem is
installed, and to default to the new ISO release that also bundles a fixed
version of the JSON gem.

This closes #103.

Signed-off-by: Daniel Pittman <daniel@rimspace.net>