(#106) Automatically generate the changelog from git commits
The manually maintained changelog file is used by the forge to inform users of
what has happened over the life of this project, but it is hard to keep
correct in the face of human error.

Since we have a complete and solid record of all the change that has happened
in the project - in the form of the git commit history - we can remove the
static version from the repository and replace it with a fully machine
generated file at build time.

This ensures the changelog is always correct, as well as rewarding the use of
good commit messages.

Closes #106.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>