(#108) Modulefile should'nt assume local directory
This change request modifies Modulefile to chdir into the root of
the project before executing git and before executing the script
./bin/git-log-to-changelog. This is necessary to get the version number
of the project and generate the CHANGELOG respectively. However, this
fails when called outside of the directory:

  fatal: Not a git repository (or any parent up to mount parent )
  Stopping at filesystem boundary (GIT_DISCOVERY_ACROSS_FILESYSTEM not set).

The change is desirable to work with other projects, like librarian-puppet
in particular that fails with an unhelpful error message because of the
exception raised in the Modulefile.