(#108) Update Modulefile to work when built outside the module root
Other tools, like `puppet-librarian`, try to build the module while running
outside the root of the checkout.  This caused some problems with version
number extraction and changelog generation, since the Modulefile assumed it
ran at the top of the checkout.

This updates that to use the full path to the various scripts, ensuring that
we build correctly when invoked from the outside.  Mostly, that means using a
path relative to the Modulefile itself.

This closes #108.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>