Remove duplicate headers from generated changelog
When the HEAD of the module checkout was a final release, such as 0.6.0, the
changelog generator would emit a duplicate header.  It was treating "0.6.0" as
the development version, and then as the head of the previous release.

This was technically correct, but useless, since there was no distinction
between the two, and so the second "in development" changelog contained
no content.

This simply skips a header when the top and bottom are the same tag, since
that will always be an empty set.  That eliminates the duplicate header, but
keeps building a correct changelog when an in-flight version exists.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>