(#112) Added rubygems_update param to razor class.
This change adds the rubygems_update parameter to the razor class so
that it can be propagated to the ::ruby class.

The first motivation is that the current version of rubygems is 2.0.1
which doesn't seem to be compatible with earlier versions. This results
in lots of errors about the --include-dependencies option which is
apparently now implied by default:

  Notice: /Stage[main]/Ruby::Dev/Package[ruby-devel]/ensure: created
    Error: Execution of '/usr/bin/gem install --include-dependencies
      --no-rdoc --no-ri require_all' returned 1:
      ERROR: While executing gem ... (OptionParser::InvalidOption)
        invalid option: --include-dependencies

The second motivation is that it makes it possible to override the default
behavior on CentOS to update rubygems. This override would be desirable
for people who don't necessarily want to pull external dependencies. It
might also be desirable for people who actually want to update rubygems
on other OS families.