Install the `net-scp` gem in addition to `net-ssh`
The external script broker, recently added to Razor, adds a new dependency on
the `net-scp` gem in addition to the `net-ssh` library; this extends the later
with an integration of scp file transfer.

This updates the Ruby manifests to install both gems, which ensures that
current Razor continues to run correctly.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>