Make `tar` installation of stable releases the default
This moves away from the previous, legacy, default of installing from git
head, in favour of installation from the stable release tar file available on
https://downloads.puppetlabs.com/razor

Like the packages, this will refuse to install over the top of an existing git
deployment; if `/opt/razor/.git` exists the Puppet module will refuse to run.
Users should manually remove the directory - or otherwise resolve
the conflict.

It is safe, and supported, to simply delete the `.git` directory, and then
install from tar over the existing git deployment.  There should not be any
data loss from this.

This closes #124.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>