Revert "Make `tar` installation of stable releases the default"
This reverts commit 9ea9092c300df325bb3eb1867281442a109b0fd0.

It turns out that the approach I took to solving the problem, while
attractive, is secretly an attractive nuisance.

It makes it impossible to actually test anything but the released version of
the code and, in the nature of hitting commit, also called a couple of people
out to explain why it made their lives miserable also.

This brings us back to the previous state, from which I will proceed forward
to add tarball support again, and to integrate it in a fashion that doesn't
lead to quite the same set of failures.