(#133) Razor image provider should not check for errors like other providers
This change reverts checking for errors in the image provider because
the image action of the razor command does not return PSON unlike the
other actions. Instead, the image provider now returns the output of
the razor command and always succeeds, like before.

Ideally, the image provider should look at the return value and raise a
Puppet::Error when failing to add an image. However, the razor command
always seems to return 0 and the output is awkward to parse. So, the
changes to the image provider should be reverted for now until the Razor
project is updated accordingly.

Closes #133