(#136) Remove global nodejs `express` package installation
The Razor module installed two versions of the nodejs `express` package, a
toolkit for writing your own web server on top of a basic HTTP accept loop.

One was installed in the Razor directory, pinned to an older version, and
fairly boringly functional.  This is the one that is actually used by Razor.

The other was a globally installed "latest" version, which was not actually
used, was kind of incompatible with the way Razor build its CGI interface, and
was incompatible with older versions of nodejs on, say, Ubuntu 12.04.

This just removes the global version: it served no actual purpose, other than
violating the nodejs community standards (throw stuff in your private
directory, no shared library code), and making Razor more apparently complex.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>