Move to `razor-server` rather than Ruby/NodeJS Razor
This is a major, incompatible update: the razor module will now install the
new `razor-server` implementation of Razor, rather than the original
implementation.  This brings a new and disparate set of dependencies.

This module makes no effort to uninstall the older Razor, aiming instead to
provide a fresh implementation.  Migration is not currently supported either.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>