Create `production.log` to reduce odds of permissions problems
One of the earliest steps that needs to be run when deploying Razor is to
migrate the database.  This is typically done as root, which means that it
will create the Sequel/Sinatra log ... as root.

That, in turn, caused a failure when starting the application as a less
privileged user for the first time.

This commit doesn't fix the problem, as such, but it does create the log file
during installation, and ensures that the permissions remain correct.

This means that we *should* be less likely to encounter this in practice.

Eventually we need a better solution inside razor-server, but for now this
should ensure first installations are smooth.

Thanks to Lucas Harms <lmickh@gmail.com> for helping with this.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>