Bind to wildcard, not loopback, when starting razor-server
Unfortunately, during testing I missed that my init script was binding
exclusively to the loopback address -- my hosts file had the hostname mapped
to loopback, so it "worked for me".

This fixes that, by explicitly binding to the wildcard address when starting
the `standalone.sh`.  How embarrassing.

Thanks to Lucas Harms <lmickh@gmail.com> for helping with this.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>