Remove obsolete test code
This test code was associated with the older version of the Razor module, and
as is the nature of "bench" testing system configuration, was dominated by
useless assertions about how the module was structured.

In the rewrite the specs and other tests became essentially useless, and
rather than rewriting the testing to assert that the manifests contained what
they contained, they were ignored.

Sadly, I didn't back that up by removing the manifest content, which led to
user confusion.  So sorry to have done that to y'all.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>