Added an ISC DHCPd configuration example
This adds an example configuration file for the ISC DHCPd that will make it
work smoothly with razor-server.  Since we don't automatically configure DHCP
operation, this should ease integration for some folks.

This is not an example of using Puppet to configure the system, which is left
as an example for the reader.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>