Revert Puppet file mode management
This reverts the explicit management of file modes through Puppet; that was
added in addition to setting the umask to our expected value during unpacking
to try and work around sites with more restrictive than "default"
umask values.

Unfortunately, it doesn't compass the full complexity of permissions; we need
to have executable permissions on various files in JBoss, etc, that I don't
want to hand-maintain lists of just to work around a problem solved at
the source.

The umask changes are retained, but the permission management changes backed out.

Signed-off-by: Daniel Pittman <daniel@puppetlabs.com>